+++ src/components/inventory/Inventory.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { Container, Row, Col } from "react-bootstrap";
 
 const Inventory = () => {
   return (
