+++ src/components/landing-room/landing-room-images/CroppedDoorFinal.png
