+++ src/components/landing-room/landing-room-images/CroppedJarFinal.png
