+++ src/components/landing-room/landing-room-images/DoorOrangeCropped.png
