+++ src/components/landing-room/landing-room-images/JarCroppedFinal.png
