+++ src/components/landing-room/landing-room-images/Tooth.png
