+++ src/components/landing-room/landing-room-images/basic-mirror.png
