+++ src/components/landing-room/landing-room-images/highlighted-mirror.png
